Add pull request template (#1)
# Pull Request

## Description

This change adds a pull request template to the repository. The template includes a section for describing the changes made, fixing issues, providing motivation and context, and listing any dependencies required for the change. This standardised format will help streamline the pull request process and ensure that all necessary information is provided when submitting changes.