Add GitHub labels and sync workflow (#29)
# Pull Request

## Description

This change introduces a new label configuration system and an automated workflow to synchronise these labels with the repository.

A new file `.github/other-configurations/labels.yml` has been added, which defines a set of default GitHub labels, language-specific labels, and size labels. This configuration includes colour codes, names, and descriptions for each label.

Additionally, a new GitHub Actions workflow `.github/workflows/sync-labels.yml` has been created. This workflow automatically syncs the labels defined in the YAML file with the repository's labels whenever changes are pushed to the `main` branch and affect the labels configuration file.

The workflow uses the `micnncim/action-label-syncer` action to perform the label synchronization, ensuring that the repository's labels are always up-to-date with the configuration file.

This change will streamline label management, provide consistency across the project, and automate the process of updating labels when needed.