Add Code of Conduct (#81)
# Pull Request

## Description

This pull request includes the addition of a new Code of Conduct
document to the project. The document outlines the community's
standards, enforcement responsibilities, and guidelines for addressing
unacceptable behavior.

Key changes:

* Added a comprehensive Code of Conduct based on the Contributor
Covenant, version 2.1, to ensure a harassment-free experience for all
community members.

Details include:

* Our Pledge: Commitment to an open, welcoming, diverse, inclusive, and
healthy community.
* Our Standards: Examples of acceptable and unacceptable behaviors.
* Enforcement Responsibilities: Roles and actions of community leaders
in enforcing the Code of Conduct.
* Scope: Applicability of the Code of Conduct within community spaces
and public representations.
* Enforcement: Procedures for reporting and addressing violations,
including enforcement guidelines with varying levels of consequences.

Fixes #82