Refactor Zizmor Action (#91)
# Pull Request

## Description

This pull request includes updates to the `code-checks` workflow and
changes to the `pyproject.toml` file to improve dependency management
and configuration settings. The most important changes include replacing
the `uv` setup with a custom action, updating dependencies, and refining
configuration settings.

Workflow improvements:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L187-R190):
Replaced the `uv` setup with a custom action `setup-python-dependencies`
and changed the command to use `poetry` for running `zizmor`.

Dependency updates:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L10-R14):
Updated `ruff` to version `0.8.6` and added `zizmor` version `1.0.0` to
the development dependencies.

Configuration refinements:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L57-R62):
Corrected the regular expression for dummy variables to ensure proper
matching.
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L34-L38):
Removed comments and adjusted the format for better readability and
consistency.