Rename GitHub Actions Steps (#105)
# Pull Request

## Description

This pull request includes changes to the `jobs:` section in two
workflow files to improve the clarity of the step names. The most
important changes are the renaming of the `Checkout` step to `Checkout
Repository` for consistency and better readability.

Changes in `.github/workflows/code-checks.yml`:

* Renamed the `Checkout` step to `Checkout Repository` in multiple job
definitions to improve clarity.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L22-R22)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L50-R50)
[[3]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L85-R85)
[[4]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L144-R144)
[[5]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L164-R164)

Changes in `.github/workflows/sync-labels.yml`:

* Renamed the `Checkout` step to `Checkout Repository` for consistency.