Refactor Git Hooks (#119)
# Pull Request

## Description

This pull request includes several changes to improve the pre-commit
hooks and add new checks for Python linting and formatting. The most
important changes include the addition of new Ruff checks, significant
modifications to the `pre-commit` script, and the removal of individual
pre-commit scripts.

### Enhancements to pre-commit hooks:

*
[`githooks/pre-commit`](diffhunk://#diff-d5fe9d324948511c959e14d0aa311bb057ce7236626bb86fa51351dc26845a4dL4-R40):
Replaced the loop running individual scripts with a series of commands
to run `clean`, `gitleaks-detect`, `prettier-check`, `format-check`, and
`ruff-check` using the `just` command. Added checks to ensure each
command passes, and provided warnings and exits if any check fails.

### Addition of Ruff checks:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR55-R58):
Added a new `ruff-checks` target to run `ruff-lint` and `ruff-format`
checks.

### Removal of individual pre-commit scripts:

*
[`githooks/pre-commit-scripts/cleanup.sh`](diffhunk://#diff-71c807d4b616e21e935714d34b5f1633a94c3d4aa38e899d962bafbe08f58c05L1-L4):
Removed the script that was previously responsible for running `just
clean`.
*
[`githooks/pre-commit-scripts/gitleaks.sh`](diffhunk://#diff-b488692093f21df73d5c8d67f3a6df3273b7199a34ca84e9c9e1a115b20df271L1-L4):
Removed the script that was previously responsible for running `gitleaks
detect`.
*
[`githooks/pre-commit-scripts/justfile.sh`](diffhunk://#diff-fcd01eccc95d37bac6ab073a69ec60e574b754774360a13f31e1a39466668fc9L1-L4):
Removed the script that was previously responsible for running `just
format`.
*
[`githooks/pre-commit-scripts/prettier.sh`](diffhunk://#diff-1e2ab1eef74462625d345d9811c6cd8e51996878583f4a9878ad8d62ec7d014dL1-L4):
Removed the script that was previously responsible for running `just
prettier-format`.
*
[`githooks/pre-commit-scripts/python.sh`](diffhunk://#diff-f13793c94aec64ea3a3f3f3bcc4f045eeb9d66b5d895396175dd6f49ae3302b1L1-L8):
Removed the script that was previously responsible for running
Python-related checks and fixes.

### Minor corrections:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL46-R46):
Corrected a typo in the comment for Ruff linting and formatting.

Fixes #118