Fix Git Hooks (#122)
# Pull Request

## Description

This pull request includes several changes aimed at improving code
quality and updating dependencies. The most important changes include
adding a new gitleaks detection command, updating the pre-commit hook
for ruff checks, and bumping the required version of a dependency.

### Improvements to code quality:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR107-R113):
Added a new `gitleaks-detect` command to detect potential security
issues in the codebase.
*
[`githooks/pre-commit`](diffhunk://#diff-d5fe9d324948511c959e14d0aa311bb057ce7236626bb86fa51351dc26845a4dL33-R34):
Updated the pre-commit hook to use `ruff-checks` instead of `ruff-check`
to ensure consistency with the command naming.

### Dependency updates:

*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L19-R19):
Updated the `required-version` for the `uv` tool from `0.6.2` to `0.6.3`
to include the latest fixes and improvements.