Fix Git Hooks Clean Command (#125)
# Pull Request

## Description

This pull request includes changes to streamline the execution of
commands in `Justfile` and `githooks/pre-commit`. The most important
changes are:

Improvements to `Justfile`:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL52-R52):
Combined the `ruff-format-fix` and `ruff-lint-fix` commands into a
single line in the `ruff-fix` target to simplify the command execution.

Enhancements to `githooks/pre-commit`:

*
[`githooks/pre-commit`](diffhunk://#diff-d5fe9d324948511c959e14d0aa311bb057ce7236626bb86fa51351dc26845a4dL6-R6):
Added the `-q` flag to the `just clean` command to suppress unnecessary
output during the pre-commit hook execution.