Update SuperLinter Permissions (#129)
# Pull Request

## Description

This pull request includes an update to the linting configuration in the
GitHub Actions workflow. The change updates the version of the
super-linter used and corrects the GitHub token environment variable.

* Updated the super-linter version from `v7.2.1` to `v7.3.0` in the
`.github/workflows/code-checks.yml` file.
* Corrected the GitHub token environment variable from `${{
secrets.GH_TOKEN }}` to `${{ secrets.GITHUB_TOKEN }}` in the
`.github/workflows/code-checks.yml` file.