Update SuperLinter EditorConfig Checker (#139)
# Pull Request

## Description

This pull request includes changes to improve the configuration and
workflow for the super-linter. The most important changes include
renaming a configuration file and updating a key in the JSON
configuration file.

Configuration improvements:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L5-R5):
Renamed from `.github/super-linter-configurations/.ecrc` and corrected
the key `SpacesAftertabs` to `SpacesAfterTabs`.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R36):
Added `EDITORCONFIG_FILE_NAME` to the environment variables.