Remove Git Hooks (#151)
# Pull Request

## Description

This pull request includes changes to the `githooks` directory,
specifically removing the `post-commit` and `pre-commit` hooks. The most
important changes are the complete removal of the `pre-commit` hook
script and the `post-commit` hook script.

Removal of githooks:

*
[`githooks/post-commit`](diffhunk://#diff-434d9a5655c19f7989a916bcec6e2e60d68a96eae07a53044290623181b520e7L1-L4):
Removed the entire `post-commit` hook script, which included commands to
update the git index.
*
[`githooks/pre-commit`](diffhunk://#diff-d5fe9d324948511c959e14d0aa311bb057ce7236626bb86fa51351dc26845a4dL1-L40):
Removed the entire `pre-commit` hook script, which included various
checks such as `gitleaks-detect`, `prettier-check`, `format-check`, and
`ruff-checks`.