Remove CodeQL Configuration (#155)
# Pull Request

## Description

This pull request removes a custom CodeQL configuration file and updates
the workflow to no longer reference it. The changes simplify the CodeQL
setup by relying on default configurations.

**CodeQL configuration changes:**

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beL1-L3):
Removed the custom CodeQL configuration file, which previously excluded
the `actions/unpinned-tag` query.

**Workflow updates:**

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L169):
Removed the reference to the deleted CodeQL configuration file in the
`config-file` parameter, allowing the workflow to use default
configurations.