Remove Pinact Verify Job (#163)
# Pull Request

## Description

This pull request includes updates to dependencies and workflow
configurations. The most important changes include the removal of the
`pinact-verify` job from the GitHub Actions workflow and updates to
specific dependency versions in `pyproject.toml`.

### Workflow configuration changes:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L207-L220):
Removed the `pinact-verify` job, which included steps for checking out
the repository and installing the latest version of `uv`.

### Dependency updates:
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L13-R19):
Updated `ruff` from version `0.11.6` to `0.11.7`.
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L13-R19):
Updated the `required-version` for `uv` from `0.6.14` to `0.6.17`.