Update Python Dependencies (#178)
# Pull Request

## Description

This pull request updates dependencies in the `pyproject.toml` file to
ensure compatibility and incorporate the latest features and fixes. The
most important changes include upgrading the `playwright` library,
updating testing and linting tools, and modifying the required version
of `uv`.

Dependency updates:

* Upgraded `playwright` from version `1.51.0` to `1.52.0` to include the
latest features and fixes.
* Updated `pytest-cov` from `6.1.0` to `6.1.1` and `ruff` from `0.11.7`
to `0.11.9` in the `dev` optional dependencies for improved testing and
linting capabilities.
* Updated the required version of `uv` from `0.6.17` to `0.7.3` to
ensure compatibility with the latest tools.