Update Dependabot for Reusable Workflows (#192)
# Pull Request

## Description

This pull request makes a minor adjustment to the
`.github/dependabot.yml` configuration file. The change updates the
exclusion pattern for dependencies to ensure proper matching.

Configuration updates:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L20-R20):
Modified the exclude-pattern for `JackPlowman/reusable-workflows` by
removing the trailing wildcard (`/*`) to match the entire directory
instead of specific files within it.